[TEST] Move tests to new modules
